refactor: Use InventoryBucketDefinition
The vault, mods, consumables and lost items count now come from the definition.

The bucket names also come from the defintion and so are now localised for free.
