+++ native/app/UI/MainDrawer.tsx
@@ -8,7 +8,7 @@ import { useSafeAreaInsets } from "react-native-safe-area-context";
 import { getFullProfile } from "@/app/bungie/BungieApi.ts";
 import { useGGStore } from "@/app/store/GGStore.ts";
 import { getGuardianClassType } from "@/app/utilities/Helpers.ts";
-import { LOGO_DARK, REFRESH_ICON, SEARCH_ICON } from "@/app/inventory/logic/Constants.ts";
+import { LOGO_DARK, REFRESH_ICON, SEARCH_ICON } from "@/app/utilities/Constants.ts";
 import InventoryHeader from "@/app/inventory/pages/InventoryHeader.tsx";
 import InventoryPages from "@/app/inventory/pages/InventoryPages.tsx";
 import Spinner from "@/app/UI/Spinner.tsx";
