+++ native/app/inventory/cells/ArtifactCell.tsx
@@ -3,7 +3,7 @@ import { Image } from "expo-image";
 import { StyleSheet, Text, View, TouchableOpacity } from "react-native";
 
 import type { DestinyIconData } from "@/app/inventory/logic/Types.ts";
-import { DEFAULT_OVERLAP_COLOR } from "@/app/inventory/logic/Constants.ts";
+import { DEFAULT_OVERLAP_COLOR } from "@/app/utilities/Constants.ts";
 import { ICON_SIZE, INNER_FRAME_SIZE } from "@/app/utilities/UISize.ts";
 
 type Props = {
