+++ native/app/inventory/cells/EngramCell.tsx
@@ -3,7 +3,7 @@ import { Image } from "expo-image";
 import { StyleSheet, Text, TouchableOpacity, View } from "react-native";
 
 import { INNER_FRAME_SIZE, common } from "@/app/utilities/UISize.ts";
-import { EMPTY_ENGRAM } from "@/app/inventory/logic/Constants.ts";
+import { EMPTY_ENGRAM } from "@/app/utilities/Constants.ts";
 import type { DestinyIconData } from "@/app/inventory/logic/Types.ts";
 
 type Props = {
