+++ native/app/inventory/pages/details/IconCell.tsx
@@ -2,7 +2,7 @@ import { Image } from "expo-image";
 import { StyleSheet, View } from "react-native";
 
 import { ICON_SIZE, INNER_FRAME_SIZE } from "@/app/utilities/UISize.ts";
-import { CRAFTED_OVERLAY } from "@/app/inventory/logic/Constants.ts";
+import { CRAFTED_OVERLAY } from "@/app/utilities/Constants.ts";
 import type { DestinyItem } from "@/app/inventory/logic/Types.ts";
 import { returnBorderColor } from "@/app/store/AccountInventoryLogic.ts";
